Defer including PdfHelper into ActionController::Base until it's loaded
ActionController::Base is lazily loaded by Rails and we don't want to prematurely load it, otherwise it may be loaded before initialization runs, then any configuration set in initialization is not copied over to ActionController::Base.

For example, the new Rails 5.2 config for `Rails.application.config.action_controller.default_protect_from_forgery` is assigned in an initializer, but ActionController::Base was loaded too early by the railtie, prior to running the initializers, so the value set was never set on ActionController::Base itself, thus never triggering the default protection.

Rails 3 is the first version where load hooks were introduced, and Rails 5.2 is the first version where an explicit hook for ActionController::Base was introduced.