Add note about options based on wkhtmltopdf version.
I would much prefer to label each option with the version of wkhtmltopdf that is required to use it, and possibly commands for older or newer versions, but I don't have to time to verify that and will be more difficult to maintain. Ultimately, the source of truth should be the wkhtmltopdf docs of your specific version. But this note at least warns the user about this requirement.

Fixes #478.