Add option `use_xvfb` to emulate an X server
On some systems an X server is not available, and thus a PDF generation
is not possible. This can be avoided when installing `xvfb-run` which
emulates an X server.

Enabling the `use_xvfb` option will wrap any `wkhtmltopdf` commands
around the command `xvfb-run`, in order to be able to create PDF's, even
without having an X server running.