Do not unlink HTML temp files immediately.
The reason for not wanting to unlink them is primariily for debugging. If an error occurs, the `wkhtmltopdf` command that had problems is shown as part of the error message. However, the references to the HTML temp files used in that command are no longer there so it's very difficult to reproduce or investigate further.

We want to keep these files around so that you can essentially run the nearly identical `wkhtmltopdf` command when there was an error and dig deeper into it.

Keeping these files does not keep them forever. They are, afterall, temp files and will be cleaned up automatically by the system.

To accomplish this, we use `.close` instead of `.close!`, which does not unlink the file immediately after closing it.

NOTE: We still want to use `.close!` on the generated PDF files for two reasons:

1. This generated PDF file is not important when an error has occurred and you are re-running the `wkhtmltopdf` command.

2. The generated PDF file is the largest of the files. Typically, the HTML files for header, footer and body are relatively small compared to the generated PDF file.

Fixes #948.